+++ jobs/inprogress_update/run.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 echo 'run inprogress-update'
-python src/inprogress_update/inprogress_update.py
+python inprogress_update.py
